[IFRT] Allow kReuseInput in RemapArrays if the number of input arrays is 1
`xla::ifrt::Client::RemapArrays()` allows `kReuseInput` for projection (view).
Previously, it expected the number of input arrays and output arrays is both 1.
This restriction has been relaxed to expecting only the number of input arrays
to be 1 because each of output arrays can be considered a projection of the
input arrays. In theory, the user could run N `RemapArrays()` for 1:1 remapping
to get 1:N remapping anyway to bypass the previous restriction, making the
restriction not effective.

PiperOrigin-RevId: 634485955